Add first blog post
Blog posts are located within the `/static/blogs` folder. I wanted to separate them not just to group them logically, but it will be helpful if we ever want to use Netlify CMS. Netlify expects assets to be in the static folder, and you can define an editor config for each sub-folder.

This also adds a separate template from the standard markdown pages so we can style blog posts as needed. I added a "template" frontmatter attribute to each markdown file to determine which template should be used.

Prettier went ham on several files, so there is some extra noise in this MR for an extra serving of fun.